rearrange codes
