+++ .gitignore
@@ -1,6 +1,3 @@
-# local
-data/
-
 ### Python ###
 # Byte-compiled / optimized / DLL files
 __pycache__/
