+++ affiliation_parser/__init__.py
@@ -1,3 +1,3 @@
-from .utils import download_grid_data
 from .parse import parse_affil
-from .matcher import match_affil
+# from .matcher import match_affil
+# from .utils import download_grid_data
