+++ affiliation_parser/data/department.json
@@ -1,0 +1,13 @@
+[
+    "laboratorio",
+    "laboratories",
+    "laboratory",
+    "laboratoire",
+    "institute",
+    "academic",
+    "department",
+    "division",
+    "faculty of ",
+    "genomics center",
+    "research station"
+]
