+++ affiliation_parser/data/institution.json
@@ -1,0 +1,4 @@
+[
+    "college",
+    "university"
+]
