+++ affiliation_parser/data/remove_institution.json
@@ -1,0 +1,11 @@
+[
+    "pharmacology",
+    "college of pharmacy",
+    "institute of zoology",
+    "institute of population",
+    "institute of bioinformatics",
+    "institute of plant",
+    "section for ",
+    "institute of clinical medicine",
+    "department of clinical"
+]
