+++ affiliation_parser/parse.py
@@ -1,117 +1,9 @@
 import re
-import string
-from unidecode import unidecode
+
 import numpy as np
-from .keywords import *
+from unidecode import unidecode
 
-
-def replace_institution_abbr(affil_text: str):
-    """
-    Replace abbreviation with full institution string
-    """
-    for university_list in UNIVERSITY_ABBR:
-        for university in university_list:
-            if university in affil_text:
-                affil_text = re.sub(university, university_list[0], affil_text)
-                return affil_text
-    return affil_text
-
-
-def append_institution_city(affil: str, location: str):
-    """
-    Append city to university that has multiple campuses if exist
-    """
-    for university_list in UNIVERSITY_MULTIPLE_CAMPUS:
-        if university_list[0] in affil.lower():
-            for city in university_list[1::]:
-                if city in location.lower() and not city in affil.lower():
-                    affil = affil + " " + city
-                    return affil
-    return affil
-
-
-def clean_text(affil_text: str):
-    """
-    Given affiliation text with abbreviation, clean that text
-    """
-    affil_text = affil_text.strip()
-    affil_text = re.sub("\t", " ", affil_text)
-    affil_text = re.sub("Dept. ", "Department ", affil_text)
-    affil_text = re.sub("Surg. ", "Sugery ", affil_text)
-    affil_text = re.sub("Univ. ", "University ", affil_text)
-    affil_text = affil_text[2:] if affil_text.startswith("2 ") else affil_text
-    affil_text = affil_text[3:] if affil_text.startswith("2. ") else affil_text
-    affil_text = re.sub(r"\*", " ", affil_text)
-    affil_text = re.sub(";", "", affil_text)
-    affil_text = re.sub("E-mail:", "", affil_text)
-    affil_text = re.sub("email:", "", affil_text)
-    affil_text = re.sub("P.O. Box", "", affil_text)
-    affil_text = replace_institution_abbr(affil_text)
-    return affil_text.strip()
-
-
-def find_country(location: str):
-    """
-    Find country from string
-    """
-    location_lower = location.lower()
-    for country in COUNTRY:
-        for c in country:
-            if c in location_lower:
-                return country[0]
-    return ""
-
-
-def check_country(affil_text: str):
-    """
-    Check if any states string from USA or UK
-    """
-    for country in ["UK"]:
-        if country in affil_text:
-            return "united kingdom"
-    for state in STATES:
-        if state in affil_text:
-            return "united states of america"
-    return ""
-
-
-def parse_email(affil_text: str):
-    """Find email from given string"""
-    match = re.search(r"[\w\.-]+@[\w\.-]+", affil_text)
-    if match is not None:
-        email = match.group()
-        if email[-1] == ".":
-            email = email[:-1]
-    else:
-        email = ""
-    return email
-
-
-def parse_zipcode(affil_text: str):
-    """
-    Parse zip code from given affiliation text
-    """
-    zip_code_group = ""
-    zip_code = re.search(r"(\d{5})([-])?(\d{4})?", affil_text)
-    if zip_code is None:
-        zip_code = re.search(r"(\d{3})([-])?(\d{4})?", affil_text)
-    else:
-        zip_code = ""
-    if zip_code:
-        zip_code_group = zip_code.groups()
-        zip_code_group = [p for p in zip_code_group if p is not None]
-        zip_code_group = "".join(zip_code_group)
-    return zip_code_group
-
-
-def parse_location(location):
-    """
-    Parse location and country from affiliation string
-    """
-    location = re.sub(r"\.", "", location).strip()
-    country = find_country(location)
-    dict_location = {"location": location.strip(), "country": country.strip()}
-    return dict_location
+from .parser import *
 
 
 def parse_affil(affil_text):
