+++ affiliation_parser/parser/parse_location.py
@@ -1,0 +1,20 @@
+def find_country(location: str):
+    """
+    Find country from string
+    """
+    location_lower = location.lower()
+    for country in COUNTRY:
+        for c in country:
+            if c in location_lower:
+                return country[0]
+    return ""
+
+
+def parse_location(location):
+    """
+    Parse location and country from affiliation string
+    """
+    location = re.sub(r"\.", "", location).strip()
+    country = find_country(location)
+    dict_location = {"location": location.strip(), "country": country.strip()}
+    return dict_location
